requirements/dev.txt
-r base.txt
django-debug-toolbar==2.2  # Конфигурируемая панель с информацией для отладки проекта
factory-boy==3.2.0
